Watershed segmentation for filament finding
The major update is the new filament finding method. The contour and the watershed based methods have been moved to a separate file (nstx_gpi_structure_finder.py).
