+++ analysis/flap_nstx.cfg
@@ -15,7 +15,8 @@
  Start delay = 0
  End delay = 0
 [Module NSTX_MDSPlus]
- Server = 'skylark.pppl.gov'
+ Server = 'skylark.pppl.gov:8501'
+# Server = 'localhost'
  Virtual name file = 'self'
  Cache data = True
  Cache directory = '/Users/mlampert/data'
