+++ analysis/nstx_gpi_tools.py
@@ -602,117 +602,6 @@ def subtract_photon_peak_2D(autocorr=None,     #INPUT autocorrelation metrix
     _autocorr[tuple(middle_index)]=fit[tuple(middle_index)]
     return _autocorr
 
-class FitEllipse:
-    """
-    Wrapper class for fitting an Ellipse and returning its important features.
-    It uses the least square approximation method combined with a Lagrangian
-    minimalization for the Eigenvalues of the problem.
-    Source:
-        http://nicky.vanforeest.com/misc/fitEllipse/fitEllipse.html
-        https://stackoverflow.com/questions/39693869/fitting-an-ellipse-to-a-set-of-data-points-in-python/48002645
-        Fitzgibbon, Pilu and Fischer in Fitzgibbon, A.W., Pilu, M., and Fischer R.B., Direct least squares fitting of ellipsees, 
-        Proc. of the 13th Internation Conference on Pattern Recognition, pp 253–257, Vienna, 1996
-    Rewritten as an object.
-    """
-    def __init__(self,  
-                 x=None,                                                        #The x coordinates of the input data as a numpy array
-                 y=None,                                                        #The y coordinates of the input data as a numpy array
-                 ):
-        """
-        Initializes (fits) the ellipse onto the given x and y data points
-        """
-        if x is None or y is None:
-            raise TypeError('x or y is not set.')
-        if len(x) < 6 or len(y) < 6:
-            raise ValueError('There should be 6 points defining the ellipse.')
-            
-        x = x[:,np.newaxis]
-        y = y[:,np.newaxis]
-        D = np.hstack((x*x, x*y, y*y, x, y, np.ones_like(x)))
-        S = np.dot(D.T,D)
-        C = np.zeros([6,6])
-        C[0,2] = C[2,0] = 2; C[1,1] = -1
-        E, V =  eig(np.dot(inv(S), C))
-        n = np.argmax(np.abs(E))
-        self.x=x
-        self.y=y
-        self.parameters = V[:,n]
-        
-    @property
-    def center(self):
-        """
-        Returns the center of the ellipse.
-        """
-        p=self.parameters
-        b,c,d,f,g,a = p[1]/2, p[2], p[3]/2, p[4]/2, p[5], p[0]
-        num = b*b-a*c
-        x0=(c*d-b*f)/num
-        y0=(a*f-b*d)/num
-        return np.array([x0,y0])
-    
-    @property
-    def axes_length(self):
-        """
-        Returns the minor and major axes lengths of the ellipse.
-        """
-        p=self.parameters
-        b,c,d,f,g,a = p[1]/2, p[2], p[3]/2, p[4]/2, p[5], p[0]
-        up = 2*(a*f*f + c*d*d + g*b*b - 2*b*d*f - a*c*g)
-        down1=(b*b-a*c)*((c-a)*np.sqrt(1+4*b*b/((a-c)*(a-c)))-(c+a))
-        down2=(b*b-a*c)*((a-c)*np.sqrt(1+4*b*b/((a-c)*(a-c)))-(c+a))
-        res1=np.sqrt(up/down1)
-        res2=np.sqrt(up/down2)
-        return np.array([res1, res2])
-    
-    @property
-    def angle_of_rotation(self):
-        """
-        Returns the angle of rotation compared to horizontal.
-        """
-        p=self.parameters
-        b,c,d,f,g,a = p[1]/2, p[2], p[3]/2, p[4]/2, p[5], p[0]
-        if b == 0:
-            return 0.
-        else:
-            return np.arctan(2*b/(a-c))/2
-        #There is a modification online which makes the whole fitting fail due to
-        #a wrong angle of rotation.
-#        if b == 0:
-#            if a > c:
-#                return 0
-#            else:
-#                return np.pi/2
-#        else: 
-#            if a > c:
-#                return np.arctan(2*b/(a-c))/2
-#            else:
-#                return np.pi/2 + np.arctan(2*b/(a-c))/2
-    @property
-    def size(self):
-        """
-        Returns the size of the ellipse in the x and y direction along its center.
-        Not part of the description in the source.
-        a contains the coefficients like this:
-        a[0]*x**2 + a[1]*x*y + a[2]*y**2 + a[3]*x + a[4]*y + a[5] = 0
-        The sizes are calculated as the solution of the 2nd order equation:
-        ysize=np.abs(y1-y2) where y1,y2 = y1,2(x=x0)
-        xsize=np.abs(x1-x2) where x1,x2 = x1,2(y=y0)
-        """
-        a=self.parameters
-        x0,y0=self.center
-        
-        ax=a[0]
-        ay=a[2]
-        bx=a[1]*y0+a[3]
-        by=a[1]*x0+a[4]
-        cx=a[2]*y0**2+a[4]*y0+a[5]
-        cy=a[0]*x0**2+a[3]*x0+a[5]
-        xsize=np.sqrt(bx**2-4*ax*cx)/np.abs(ax)
-        ysize=np.sqrt(by**2-4*ay*cy)/np.abs(ay)
-        if np.imag(xsize) != 0 or np.imag(xsize) !=0:
-            print('size is complex')
-            raise ValueError('')
-        return np.array([xsize,ysize])
         
 def make_plot_cursor_format(current, other):
     """    
@@ -731,377 +620,6 @@ def make_plot_cursor_format(current, other):
         return ('Left: {:<40}    Right: {:<}'
                 .format(*['({:.6f}, {:.6f})'.format(x, y) for x,y in coords]))
     return format_coord
-
-class Polygon:
-    
-    def __init__(self,
-                 x=None,
-                 y=None):
-        if ((x is not None and y is not None) and 
-            (len(x) == len(y))):
-            self.x=x
-            self.y=y
-        else:
-            raise IOError('The input x and y has to be defined and must have the same length.')
-    @property
-    def area(self):
-        """
-        Returns the area of the polygon based on the so called shoelace formula.
-        Sources: 
-            https://stackoverflow.com/questions/24467972/calculate-area-of-polygon-given-x-y-coordinates
-            https://en.wikipedia.org/wiki/Shoelace_formula
-        """
-        return 0.5*np.abs(np.dot(self.x,np.roll(self.y,1))-np.dot(self.y,np.roll(self.x,1)))
-    @property
-    def signed_area(self):
-        return 0.5*(np.dot(self.x,np.roll(self.y,1))-np.dot(self.y,np.roll(self.x,1)))
-    
-    @property
-    def centroid(self):
-        """
-        Source: https://en.wikipedia.org/wiki/Polygon
-        Programming based on the area's convention.
-        """
-        x_center=1/(6*self.signed_area) * np.dot(self.x+np.roll(self.x,1),self.x*np.roll(self.y,1)-np.roll(self.x,1)*self.y)
-        y_center=1/(6*self.signed_area) * np.dot(self.y+np.roll(self.y,1),self.x*np.roll(self.y,1)-np.roll(self.x,1)*self.y)
-        return np.asarray([x_center,y_center])
-    
-def nstx_gpi_one_frame_structure_finder(data_object=None,                       #Name of the FLAP.data_object
-                                        exp_id='*',                             #Shot number (if data_object is not used)
-                                        time=None,                              #Time when the structures need to be evaluated (when exp_id is used)
-                                        sample=None,                            #Sample number where the structures need to be evaluated (when exp_id is used)
-                                        spatial=False,                          #Calculate the results in real spatial coordinates
-                                        pixel=False,                            #Calculate the results in pixel coordinates
-                                        mfilter_range=5,                        #Range of the median filter
-                                        nlevel=80//5,                           #The number of contours to be used for the calculation (default:ysize/mfilter_range=80//5)
-                                        levels=None,                            #Contour levels from an input and not from automatic calculation
-                                        threshold_level=None,                   #Threshold level over which it is considered to be a structure
-                                                                                    #if set, the value is subtracted from the data and contours are found after that. 
-                                                                                    #Negative values are substituted with 0.
-                                        filter_struct=True,                     #Filter out the structures with less than filter_level number of contours
-                                        filter_level=None,                      #The number of contours threshold for structures filtering (default:nlevel//4)
-                                        remove_interlaced_structures=False,     #Filter out the structures which are interlaced. Only the largest structures is preserved, others are removed.
-                                        test_result=False,                      #Test the result only (plot the contour and the found structures)
-                                        test=False,                             #Test the contours and the structures before any kind of processing
-                                        save_data_for_publication=False,
-                                        ):
-    
-    """
-    The method calculates the radial and poloidal sizes of the structures
-    present in one from of the GPI image. It gathers the isosurface contour
-    coordinates and determines the structures based on certain criteria. In
-    principle no user input is necessary, the code provides a robust solution.
-    The sizes are determined by fitting an ellipse onto the contour at
-    half-height. The code returns the following list:
-        a[structure_index]={'Paths':     [list of the paths, type: matplotlib.path.Path],
-                            'Half path': [path at the half level of the structure]
-                            'Levels':    [levels of the paths, type: list],
-                            'Center':    [center of the ellipse in px,py or R,z coordinates, type: numpy.ndarray of two elements],
-                            'Size':      [size of the ellipse in x and y direction or R,z direction, type: ]numpy.ndarray of two elements,
-                            'Angle':      [angle of the ellipse compared to horizontal in radians, type: numpy.float64],
-                            'Area':      [area of the polygon at the half level],
-                            ('Ellipse':  [the entire ellipse object, returned if test_result is True, type: flap_nstx.analysis.nstx_gpi_tools.FitEllipse])
-                            }
-    """
-    
-    if type(data_object) is str:
-        data_object=flap.get_data_object_ref(data_object, exp_id=exp_id)
-        if len(data_object.data.shape) != 2:
-            raise IOError('The inpud data_object is not 2D. The method only processes 2D data.')
-            
-    if data_object is None:
-        if (exp_id is None) or ((time is None) and (sample is None)):
-            raise IOError('exp_id and time needs to be set if data_object is not set.')
-        try:
-            data_object=flap.get_data_object_ref('GPI', exp_id=exp_id)
-        except:
-            print('---- Reading GPI data ----')
-            data_object=flap.get_data('NSTX_GPI', exp_id=exp_id, name='', object_name='GPI')
-            
-        if (time is not None) and (sample is not None):
-            raise IOError('Either time or sample can be set, not both.')
-            
-        if time is not None:
-            data_object=data_object.slice_data(slicing={'Time':time})
-        if sample is not None:
-            data_object=data_object.slice_data(slicing={'Sample':sample})
-            
-    try:
-        data_object.data
-    except:
-        raise IOError('The input data object should be a flap.DataObject')
-        
-    if len(data_object.data.shape) != 2:
-        raise TypeError('The frame dataobject needs to be a 2D object without a time coordinate.')
-        
-    if pixel:
-        x_coord_name='Image x'
-        y_coord_name='Image y'
-    if spatial:
-        x_coord_name='Device R'
-        y_coord_name='Device z'
-        
-    x_coord=data_object.coordinate(x_coord_name)[0]
-    y_coord=data_object.coordinate(y_coord_name)[0]
-    data = scipy.ndimage.median_filter(data_object.data, mfilter_range)
-    if test:
-        plt.cla()
-        
-    if threshold_level is not None:
-        if data.max() < threshold_level:
-            print('The maximum of the signal doesn\'t reach the threshold level.')
-            return None
-        data = data - threshold_level
-        data[np.where(data < 0)] = 0.
-        
-    if levels is None:
-        levels=np.arange(nlevel)/(nlevel-1)*(data.max()-data.min())+data.min()
-    else:
-        nlevel=len(levels)
-        
-    try:
-        structure_contours=plt.contourf(x_coord, y_coord, data, levels=levels)
-    except:
-        plt.cla()
-        plt.close()
-        print('Failed to create the contours for the structures.')
-        return None
-    
-    if not test:
-        plt.cla()
-    structures=[]
-    one_structure={'Paths':[None], 
-                   'Levels':[None]}
-    if test:
-        print('Plotting levels')
-    else:
-        plt.close()
-    #The following lines are the core of the code. It separates the structures
-    #from each other and stores the in the structure list.
-    
-    """
-    Steps of the algorithm:
-        
-        1st step: Take the paths at the highest level and store them. These
-                  create the initial structures
-        2nd step: Take the paths at the second highest level
-            2.1 step: if either of the previous paths contain either of
-                      the paths at this level, the corresponding
-                      path is appended to the contained structure from the 
-                      previous step.
-            2.2 step: if none of the previous structures contain the contour
-                      at this level, a new structure is created.
-        3rd step: Repeat the second step until it runs out of levels.
-        4th step: Delete those structures from the list which doesn't have
-                  enough paths to be called a structure.
-                  
-    (Note: a path is a matplotlib path, a structure is a processed path)
-    """
-    for i_lev in range(len(structure_contours.collections)-1,-1,-1):
-        cur_lev_paths=structure_contours.collections[i_lev].get_paths()
-        n_paths_cur_lev=len(cur_lev_paths)
-        
-        if len(cur_lev_paths) > 0:
-            if len(structures) == 0:
-                for i_str in range(n_paths_cur_lev):
-                    structures.append(copy.deepcopy(one_structure))
-                    structures[i_str]['Paths'][0]=cur_lev_paths[i_str]
-                    structures[i_str]['Levels'][0]=levels[i_lev]
-            else:
-                for i_cur in range(n_paths_cur_lev):
-                    new_path=True
-                    cur_path=cur_lev_paths[i_cur]
-                    for j_prev in range(len(structures)):
-                        if cur_path.contains_path(structures[j_prev]['Paths'][-1]):
-                            structures[j_prev]['Paths'].append(cur_path)
-                            structures[j_prev]['Levels'].append(levels[i_lev])
-                            new_path=False
-                    if new_path:
-                        structures.append(copy.deepcopy(one_structure))
-                        structures[-1]['Paths'][0]=cur_path
-                        structures[-1]['Levels'][0]=levels[i_lev]
-                    if test:
-                        x=cur_lev_paths[i_cur].to_polygons()[0][:,0]
-                        y=cur_lev_paths[i_cur].to_polygons()[0][:,1]
-                        plt.plot(x,y)
-                        plt.axis('equal')
-                        plt.pause(0.001)
-    
-    #Cut the structures based on the filter level
-    if filter_level is None:
-        filter_level=nlevel//5
-    if filter_struct:
-        cut_structures=[]
-        for i_str in range(len(structures)):
-            if len(structures[i_str]['Levels']) > filter_level:
-                cut_structures.append(structures[i_str])
-    structures=cut_structures
-    
-    if test:
-        print('Plotting structures')
-        plt.cla()
-        plt.set_aspect(1.0)
-        for struct in structures:
-            plt.contourf(x_coord, y_coord, data, levels=levels)
-            for path in struct['Paths']:
-                x=path.to_polygons()[0][:,0]
-                y=path.to_polygons()[0][:,1]
-                plt.plot(x,y)
-            plt.pause(1)
-            plt.cla()
-            #plt.axis('equal')
-            plt.set_aspect(1.0)
-        plt.contourf(x_coord, y_coord, data, levels=levels)
-        plt.colorbar()           
-
-    #Finding the contour at the half level for each structure and
-    #calculating its properties
-    if len(structures) > 1:
-        #Finding the paths at FWHM
-        paths_at_half=[]
-        for i_str in range(len(structures)):
-            half_level=(structures[i_str]['Levels'][-1]+structures[i_str]['Levels'][0])/2.
-            ind_at_half=np.argmin(np.abs(structures[i_str]['Levels']-half_level))
-            paths_at_half.append(structures[i_str]['Paths'][ind_at_half])
-        #Process the structures which are embedded (cut the inner one)
-        if remove_interlaced_structures:
-            structures_to_be_removed=[]
-            for ind_path1 in range(len(paths_at_half)):
-                for ind_path2 in range(len(paths_at_half)):
-                    if ind_path1 != ind_path2:
-                        if paths_at_half[ind_path2].contains_path(paths_at_half[ind_path1]):
-                            structures_to_be_removed.append(ind_path1)
-            structures_to_be_removed=np.unique(structures_to_be_removed)
-            cut_structures=[]
-            for i_str in range(len(structures)):
-                if i_str not in structures_to_be_removed:
-                    cut_structures.append(structures[i_str])
-            structures=cut_structures
-        
-    #Calculate the ellipse and its properties for the half level contours    
-    for i_str in range(len(structures)):
-        str_levels=structures[i_str]['Levels']
-        half_level=(str_levels[-1]+str_levels[0])/2.
-        ind_at_half=np.argmin(np.abs(str_levels-half_level))
-        n_path=len(structures[i_str]['Levels'])
-        polygon_areas=np.zeros(n_path)
-        polygon_centroids=np.zeros([n_path,2])
-        polygon_intensities=np.zeros(n_path)
-        for i_path in range(n_path):
-            polygon=structures[i_str]['Paths'][i_path].to_polygons()
-            if polygon != []:
-                polygon=polygon[0]
-                polygon_areas[i_path]=flap_nstx.analysis.Polygon(polygon[:,0],polygon[:,1]).area
-                polygon_centroids[i_path,:]=flap_nstx.analysis.Polygon(polygon[:,0],polygon[:,1]).centroid
-            if i_path == 0:
-                polygon_intensities[i_path]=polygon_areas[i_path]*str_levels[i_path]
-            else:
-                polygon_intensities[i_path]=(polygon_areas[i_path]-polygon_areas[i_path-1])*str_levels[i_path]
-        intensity=np.sum(polygon_intensities)
-        center_of_gravity=np.asarray([np.sum(polygon_intensities*polygon_centroids[:,0])/np.sum(polygon_intensities),
-                                      np.sum(polygon_intensities*polygon_centroids[:,1])/np.sum(polygon_intensities)])
-        half_coords=structures[i_str]['Paths'][ind_at_half].to_polygons()[0]
-        half_polygon=flap_nstx.analysis.Polygon(half_coords[:,0],half_coords[:,1])
-        
-        structures[i_str]['Half path']=structures[i_str]['Paths'][ind_at_half]
-        structures[i_str]['Half level']=half_level
-        structures[i_str]['Centroid']=half_polygon.centroid
-        structures[i_str]['Area']=half_polygon.area
-        structures[i_str]['Intensity']=intensity
-        structures[i_str]['Center of gravity']=center_of_gravity
-        try:
-            ellipse=flap_nstx.analysis.FitEllipse(half_coords[:,0],half_coords[:,1])
-            structures[i_str]['Center']=ellipse.center
-            size=ellipse.size
-            structures[i_str]['Size']=size
-            structures[i_str]['Angle']=ellipse.angle_of_rotation
-            structures[i_str]['Elongation']=(size[0]-size[1])/(size[0]+size[1])
-            
-            if test_result or test:
-                structures[i_str]['Ellipse']=ellipse
-            else:
-                structures[i_str]['Ellipse']=None
-        except:
-            print('Ellipse fitting failed.')
-
-            structures[i_str]['Center']=None
-            structures[i_str]['Size']=None
-            structures[i_str]['Angle']=None
-            structures[i_str]['Elongation']=None
-            
-            structures[i_str]['Ellipse']=None
-            
-    fitted_structures=[]        
-    for i_str in range(len(structures)):
-        if structures[i_str]['Size'] is not None:
-            fitted_structures.append(structures[i_str])
-            
-    structures=fitted_structures
-    if test_result:
-        
-        fig,ax=plt.subplots(figsize=(8.5/2.54, 8.5/2.54/1.62))
-        ax.set_aspect(1.0)
-        plt.contourf(x_coord, y_coord, data, levels=levels)
-        plt.colorbar()  
-        if len(structures) > 0:
-            #Parametric reproduction of the Ellipse
-            R=np.arange(0,2*np.pi,0.01)
-            for i_structure in range(len(structures)):
-                structure=structures[i_structure]
-                if structure['Half path'] is not None:
-                    phi=structure['Angle']
-                    a,b=structure['Ellipse'].axes_length
-                    x=structure['Half path'].to_polygons()[0][:,0]
-                    y=structure['Half path'].to_polygons()[0][:,1]
-                    xx = structure['Center'][0] + \
-                         a*np.cos(R)*np.cos(phi) - \
-                         b*np.sin(R)*np.sin(phi)
-                    yy = structure['Center'][1] + \
-                         a*np.cos(R)*np.sin(phi) + \
-                         b*np.sin(R)*np.cos(phi)
-                    plt.plot(x,y)    #Plot the half path polygon
-                    plt.plot(xx,yy)  #Plot the ellipse
-                    plt.scatter(structure['Centroid'][0],
-                                structure['Centroid'][1], color='yellow')
-                    plt.scatter(structure['Center of gravity'][0],
-                                structure['Center of gravity'][1], color='red')
-                    if save_data_for_publication:
-                        exp_id=data_object.exp_id
-                        time=data_object.coordinate('Time')[0][0,0]
-                        wd=flap.config.get_all_section('Module NSTX_GPI')['Working directory']
-                        filename=wd+'/'+str(exp_id)+'_'+str(time)+'_half_path_no.'+str(i_structure)+'.txt'
-                        file1=open(filename, 'w+')
-                        for i in range(len(x)):
-                            file1.write(str(x[i])+'\t'+str(y[i])+'\n')
-                        file1.close()
-                        
-                        filename=wd+'/'+str(exp_id)+'_'+str(time)+'_fit_ellipse_no.'+str(i_structure)+'.txt'
-                        file1=open(filename, 'w+')
-                        for i in range(len(xx)):
-                            file1.write(str(xx[i])+'\t'+str(yy[i])+'\n')
-                        file1.close()
-                        
-                plt.xlabel('Image x')
-                plt.ylabel('Image y')
-                plt.title(str(exp_id)+' @ '+str(data_object.coordinate('Time')[0][0,0]))
-                plt.show()
-                plt.pause(0.1)
-                
-        if save_data_for_publication:
-            exp_id=data_object.exp_id
-            time=data_object.coordinate('Time')[0][0,0]
-            wd=flap.config.get_all_section('Module NSTX_GPI')['Working directory']
-            filename=wd+'/'+str(exp_id)+'_'+str(time)+'_raw_data.txt'
-            file1=open(filename, 'w+')
-            for i in range(len(data[0,:])):
-                string=''
-                for j in range(len(data[:,0])):
-                    string+=str(data[j,i])+'\t'
-                string+='\n'
-                file1.write(string)
-            file1.close()
-                
-    return structures
 
 def signal_windowed_avg_err(x,windowsize):
     """
